윤형빈 : 업데이트

+++ GameEngineContents/Bar.cpp
@@ -58,16 +58,16 @@ if (BarTime > 1.0f)
 		{
 			CommentTextureRenderer* Renderer = CreateComponent<CommentTextureRenderer>();
 			Renderer->SetTexture("Color2.png");
-			Renderer->GetTransform().SetLocalPosition(float4(343.5f, 168.5 + 27 * (GlobalContentsValue::Stress) * 0.01f * 0.5f,0));
-			Renderer->GetTransform().SetLocalScale(float4(3, 27 * (GlobalContentsValue::Stress) * 0.01f));
+			Renderer->GetTransform().SetLocalPosition(float4(343.5f, 168.5f + 27.0f * static_cast<float>(GlobalContentsValue::Stress) * 0.01f * 0.5f,0.0f));
+			Renderer->GetTransform().SetLocalScale(float4(3.0f, 27.0f * static_cast<float>(GlobalContentsValue::Stress) * 0.01f));
 			Renderer->Clip.SizeX = 760.0f;
 			StressStack.push_back(Renderer);
 		}
 		{
 			CommentTextureRenderer* Renderer = CreateComponent<CommentTextureRenderer>();
 			Renderer->SetTexture("Color2.png");
-			Renderer->GetTransform().SetLocalPosition(float4(343.5f, 132.5f + 27 * (GlobalContentsValue::Affection) * 0.01f * 0.5f,0));
-			Renderer->GetTransform().SetLocalScale(float4(3, 27 * (GlobalContentsValue::Affection) * 0.01f));
+			Renderer->GetTransform().SetLocalPosition(float4(343.5f, 132.5f + 27.0f * static_cast<float>(GlobalContentsValue::Affection) * 0.01f * 0.5f,0.0f));
+			Renderer->GetTransform().SetLocalScale(float4(3.0f, 27.0f * static_cast<float>(GlobalContentsValue::Affection) * 0.01f));
 			Renderer->Clip.SizeX = 760.0f;
 			AffectionStack.push_back(Renderer);
 		}
@@ -85,8 +85,8 @@ void Bar::BarMove()
 		(*StartIter)->GetTransform().SetLocalMove(float4(-3, 0, 0));
 	}
 	StartIter = MentalStack.begin() + BarCount;
-	(*StartIter)->GetTransform().SetLocalPosition(float4(343.5f, 97.5f + 27 * (GlobalContentsValue::Mental) * 0.01f * 0.5f, 0));
-	(*StartIter)->GetTransform().SetLocalScale(float4(3.0f, 27.0f * (GlobalContentsValue::Mental) * 0.01f));
+	(*StartIter)->GetTransform().SetLocalPosition(float4(343.5f, 97.5f + 27.0f * static_cast<float>(GlobalContentsValue::Mental) * 0.01f * 0.5f, 0.0f));
+	(*StartIter)->GetTransform().SetLocalScale(float4(3.0f, 27.0f * static_cast<float>(GlobalContentsValue::Mental) * 0.01f));
 
 	StartIter = StressStack.begin();
 	EndIter = StressStack.end();
@@ -95,8 +95,8 @@ void Bar::BarMove()
 		(*StartIter)->GetTransform().SetLocalMove(float4(-3, 0, 0));
 	}
 	StartIter = StressStack.begin() + BarCount;
-	(*StartIter)->GetTransform().SetLocalPosition(float4(343.5f, 168.5 + 27 * (GlobalContentsValue::Stress) * 0.01f * 0.5f, 0));
-	(*StartIter)->GetTransform().SetLocalScale(float4(3.0f, 27.0f * (GlobalContentsValue::Stress) * 0.01f));
+	(*StartIter)->GetTransform().SetLocalPosition(float4(343.5f, 168.5f + 27.0f * static_cast<float>(GlobalContentsValue::Stress) * 0.01f * 0.5f, 0.0f));
+	(*StartIter)->GetTransform().SetLocalScale(float4(3.0f, 27.0f * static_cast<float>(GlobalContentsValue::Stress) * 0.01f));
 
 	StartIter = AffectionStack.begin();
 	EndIter = AffectionStack.end();
@@ -105,8 +105,8 @@ void Bar::BarMove()
 		(*StartIter)->GetTransform().SetLocalMove(float4(-3, 0, 0));
 	}
 	StartIter = AffectionStack.begin() + BarCount;
-	(*StartIter)->GetTransform().SetLocalPosition(float4(343.5f, 132.5 + 27 * (GlobalContentsValue::Affection) * 0.01f * 0.5f, 0));
-	(*StartIter)->GetTransform().SetLocalScale(float4(3.0f, 27.0f * (GlobalContentsValue::Affection) * 0.01f));
+	(*StartIter)->GetTransform().SetLocalPosition(float4(343.5f, 132.5f + 27.0f * static_cast<float>(GlobalContentsValue::Affection) * 0.01f * 0.5f, 0.0f));
+	(*StartIter)->GetTransform().SetLocalScale(float4(3.0f, 27.0f * static_cast<float>(GlobalContentsValue::Affection) * 0.01f));
